Move logging module to the runtime package
