+++ runtime/log.go
@@ -1,4 +1,4 @@
-package log
+package runtime
 
 import (
 	"fmt"
